Bug Fix - Creating Event in the Correct Day
Problem: When I created a new event, the form opened only in the last day.

Cause: I think when I moved the responsibility of the event-form to the context I lost the reference day.

Solution: I created a new state call event day in context and change it when the user clicks to create an event.